==> notebooks/01_get_grb_GRBWeb.ipynb <==
Run this on `gatech-hive` \
~Running in pre-compiled env `/cvmfs/icecube.opensciencegrid.org/py3-v4.1.0/RHEL_7_x86_64/metaprojects/combo/stable/env-shell.sh` on cobalt06~ \

Get grb list from GRBWeb \
Start date and end date are aligned with ~GRECO v2.4 data~ GRECO v2.10 data

import numpy as np
import sqlite3
import pandas
import os

# Download the SQLite file from the GRBweb webpage
# os.system("wget -O /home/cjchen/2020-2021/Spring2021/greco_grb/data/grbweb/GRBweb2.sqlite https://icecube.wisc.edu/~grbweb_public/GRBweb2.sqlite")
os.system("wget -O /storage/home/hhive1/cchen641/data/icecube/data/greco_grb/data/grbweb/GRBweb2.sqlite https://icecube.wisc.edu/~grbweb_public/GRBweb2.sqlite")


# db = sqlite3.connect('/home/cjchen/2020-2021/Spring2021/greco_grb/data/grbweb/GRBweb2.sqlite')
db = sqlite3.connect('/storage/home/hhive1/cchen641/data/icecube/data/greco_grb/data/grbweb/GRBweb2.sqlite')
db

# Print the names of all the tables
table_names = pandas.read_sql_query("SELECT * from sqlite_sequence", db)
table_names


# Get numpy arrays containing the right ascension, declination, mjd, ... of the entries in the Summary table
Summary_table = pandas.read_sql_query("SELECT * from Summary", db)
Summary_table.head()

keys = Summary_table.keys()
print("Keys in the 'Summary' table: {}\n".format(list(keys)))
RA  = np.array(Summary_table.ra)   # right ascension
DEC = np.array(Summary_table.decl) # declination
MJD = np.array(Summary_table.mjd)  # modified julian date
print(" RA: {}".format(RA[:3]))
print(" DEC: {}".format(DEC[:3]))
print(" MJD: {}".format(MJD[:3]))

Summary_table.shape

## GRECO v2.4
`/data/user/mlarson/combo_r129072/scripts/greco_online/skylab_dataset/greco_online/current` \
Data 2012 - 2018 \
from MJD `56043.42313825331` to `58648.87564821623` \
Or `2012-04-26 10:09:19.145 UTC` ~ `2019-06-14 21:00:56.006 UTC` \
No boundary constraint here.

## GRECO v2.10
`/data/user/mlarson/combo_r129072/scripts/greco_online/skylab_dataset/greco_online/candidate-002-p10` \
Data 2012 - 2019 \
GRL starts from MJD `56043.42452709535` and ends at `58998.82821445378` \
Or `2012-04-26 10:11:19.141 UTC` ~ `2020-05-29 19:52:37.729 UTC` \
Boundary constraints:
`t_center > 56043.42452709535 + 500./86400` and `t_center < 58998.82821445378 - 500./86400`

# test1 = np.load("/data/user/mlarson/combo_r129072/scripts/greco_online/skylab_dataset/greco_online/candidate-002-p10/GRL/IC86_2012.data.npy")
# test2 = np.load("/data/user/mlarson/combo_r129072/scripts/greco_online/skylab_dataset/greco_online/candidate-002-p10/GRL/IC86_2019.data.npy")
test1 = np.load("/storage/home/hhive1/cchen641/data/icecube/data/analyses/online_GRECO/candidate-002-p10/GRL/IC86_2012.data.npy")
test2 = np.load("/storage/home/hhive1/cchen641/data/icecube/data/analyses/online_GRECO/candidate-002-p10/GRL/IC86_2019.data.npy")

print(np.min(test1['start']))
print(np.max(test2['stop']))


if np.isnan(np.sum(MJD)):
    print("We have null MJD")
else:
    print("All bursts have MJD")

All angles in degrees \
All times in days

Paul: for IPN GRBs in the database, 1x long dimension is recorded as pos_error_source

What I what is radius, so from those GRBs I will multiply the pos_error with 0.5x

query = """
SELECT 
  -- name
  GRB_name AS grb_name, 
  GRB_name_Fermi AS grb_name_GBM, 
  -- space
  ra, 
  decl AS dec, 
  CASE pos_error_source WHEN 'IPN' THEN 0.5 * pos_error ELSE pos_error END source_uncertainty, 
  -- time
  mjd AS mjd_start, 
  T100 / 86400 AS t100, 
  T90 / 86400 AS t90, 
  -- satellite 
  GBM_located AS gbm, 
  ra_source, 
  decl_source, 
  pos_error_source, 
  -- other 
  fluence, 
  redshift, 
  -- engineered 
  (mjd + 0.5 * T100 / 86400) AS t_center 
FROM 
  Summary 
WHERE 
  (T100 IS NOT NULL) 
  And (
    t_center BETWEEN 56043.43031413238 
    And 58998.822427416744
  ) 
ORDER BY 
  grb_name DESC
"""
grb_table_2297 = pandas.read_sql_query(query, db)
grb_table_2297.head()

grb_table_2297.shape

type(grb_table_2297)

# grb_table.to_pickle("/home/cjchen/2020-2021/Spring2021/greco_grb/data/grbweb/grb_grbweb.pkl")
# grb_table_2297.to_pickle("/storage/home/hhive1/cchen641/data/icecube/data/greco_grb/data/grbweb/grb_grbweb_2297.pkl")
