==> notebooks/16_misc.ipynb <==
# mjd to utc

import numpy as np
import pandas as pd
import pickle
import healpy as hp
import histlite as hl
import os
from glob import glob
import astropy as ap
from astropy.time import Time


result01 = pd.read_pickle("/storage/home/hhive1/cchen641/icecube/greco_grb/data/csky_output/unblind/result01.pkl")


times = result01.mjd_start
t = Time(times, format='mjd')

def format_utc(utc):
    # python can't have \b --> escape character e.g. \n
    date, time = utc.split()
    return time
utc_array = np.array(list(map(format_utc, t.iso.astype('U'))))

utc_array

result02 = result01.copy()
result02['mjd_start_utc'] = utc_array

result02.to_pickle("/storage/home/hhive1/cchen641/icecube/greco_grb/data/csky_output/unblind/result02.pkl")


# GRECO v2.10 and GFU v2.6 Overlap

import sys
sys.path.append('../../')
import csky as cy
from greco_grb.scripts import SETTING
paths = SETTING.PATH()
print(paths)
USER = paths.USER
ICDATA_DIR = paths.ICDATA_DIR
DATA_DIR = paths.DATA_DIR
ANA_DIR = paths.ANA_DIR
from greco_grb.scripts.utils import *

print("\n===== Setting up csky =====\n")
data_dir = ICDATA_DIR
data_filenames = sorted(glob(data_dir + '/IC86_20*.data.npy'))
sig_filenames = sorted(glob(data_dir + '/IC86_2012.nu*_merged.npy'))
grl_filenames = sorted(glob(data_dir + '/GRL/IC86_20*.data.npy'))

################ energy lower bound #############
min_log_e = np.log10(10)
#################################################
bins_sindec = np.linspace(-1, 1, 25+1)  
bins_logenergy = np.linspace(min_log_e, 4, 25+1)

data = [np.load(data_filename) for data_filename in data_filenames]
data = np.hstack(data)
sig = [np.load(sig_filename) for sig_filename in sig_filenames]
sig = np.hstack(sig)
grl = [np.load(grl_filename) for grl_filename in grl_filenames]
grl = np.hstack(grl)
if min_log_e is not None:
    data_mask = data['logE'] > min_log_e
    data = data[data_mask]
    sig_mask = sig['logE'] > min_log_e
    sig = sig[sig_mask]
    
dataset_spec = cy.selections.CustomDataSpecs.CustomDataSpec(data, sig, np.sum(grl['livetime']),
                                                     sindec_bins=bins_sindec,
                                                     logenergy_bins=bins_logenergy,
                                                     grl=grl, key='greco_v2.10', cascades=True)

ANA_DIR = cy.utils.ensure_dir(ANA_DIR)
# on OSG
# ana_dir = "./"
ana = cy.get_analysis(cy.selections.repo
                      , dataset_spec
                      , dir=ANA_DIR
                      , load_sig=True)  # to save memory: use False


len(ana[0].data.event)

print("\n===== Setting up csky =====\n")
data_dir = ICDATA_DIR
data_filenames = sorted(glob(data_dir + '/IC86_20*.data.npy'))
sig_filenames = sorted(glob(data_dir + '/IC86_2012.nu*_merged.npy'))
grl_filenames = sorted(glob(data_dir + '/GRL/IC86_20*.data.npy'))

################ energy lower bound #############
min_log_e = np.log10(1e-6)
#################################################
bins_sindec = np.linspace(-1, 1, 25+1)  
bins_logenergy = np.linspace(min_log_e, 8, 25+1)

data = [np.load(data_filename) for data_filename in data_filenames]
data = np.hstack(data)
sig = [np.load(sig_filename) for sig_filename in sig_filenames]
sig = np.hstack(sig)
grl = [np.load(grl_filename) for grl_filename in grl_filenames]
grl = np.hstack(grl)
if min_log_e is not None:
    data_mask = data['logE'] > min_log_e
    data = data[data_mask]
    sig_mask = sig['logE'] > min_log_e
    sig = sig[sig_mask]
    
dataset_spec = cy.selections.CustomDataSpecs.CustomDataSpec(data, sig, np.sum(grl['livetime']),
                                                     sindec_bins=bins_sindec,
                                                     logenergy_bins=bins_logenergy,
                                                     grl=grl, key='greco_v2.10', cascades=True)

ANA_DIR = cy.utils.ensure_dir(ANA_DIR)
# on OSG
# ana_dir = "./"
ana = cy.get_analysis(cy.selections.repo
                      , dataset_spec
                      , dir=ANA_DIR
                      , load_sig=True)  # to save memory: use False


data.size

print("\n===== Setting up csky =====\n")
gfu_data_dir = "/storage/home/hhive1/cchen641/data/icecube/data/analyses/gfu/version-002-p06/"
data_filenames = sorted(glob(gfu_data_dir + '/IC86_20*data.npy'))
sig_filenames = sorted(glob(gfu_data_dir + '/IC86_*_MC.npy'))
grl_filenames = sorted(glob(gfu_data_dir + '/GRL/IC86_20*data.npy'))

################ energy lower bound #############
# min_log_e = np.log10(1e-6)
#################################################
bins_sindec = np.unique(np.concatenate([
             np.linspace(-1, -0.93, 4 + 1),
             np.linspace(-0.93, -0.3, 10 + 1),
             np.linspace(-0.3, 0.05, 9 + 1),
             np.linspace(0.05, 1, 18 + 1) ])) 
bins_logenergy = np.arange(1, 9.5 + 0.01, 0.125)

data = [np.load(data_filename) for data_filename in data_filenames]
data = np.hstack(data)
sig = [np.load(sig_filename) for sig_filename in sig_filenames]
sig = np.hstack(sig)
grl = [np.load(grl_filename) for grl_filename in grl_filenames]
grl = np.hstack(grl)
# if min_log_e is not None:
#     data_mask = data['logE'] > min_log_e
#     data = data[data_mask]
#     sig_mask = sig['logE'] > min_log_e
#     sig = sig[sig_mask]
    
dataset_spec = cy.selections.CustomDataSpecs.CustomDataSpec(data, sig, np.sum(grl['livetime']),
                                                     sindec_bins=bins_sindec,
                                                     logenergy_bins=bins_logenergy,
                                                     grl=grl, key='gfu_v2.6', cascades=False)

ANA_DIR = cy.utils.ensure_dir(ANA_DIR)
# on OSG
# ana_dir = "./"
gfu_ana = cy.get_analysis(cy.selections.repo
                      , dataset_spec
                      , dir=ANA_DIR
                      , load_sig=True)  # to save memory: use False


data.size

gfu_ana[0].data

event_ids = set(ana[0].data.event)


len(ana[0].data.event)

count = 0
for event_id in gfu_ana[0].data.event:
    if(event_id in event_ids):
#         print(event_id)
        count += 1
print(count)

47811/1010151, 47811/1532457
