==> notebooks/17_stacked_TS.ipynb <==
# Only import neccessary packages
import pandas as pd
import sys
sys.path.append('../../')
import csky as cy
from greco_grb.scripts import SETTING
paths = SETTING.PATH()
print(paths)
USER = paths.USER
ICDATA_DIR = paths.ICDATA_DIR
DATA_DIR = paths.DATA_DIR
ANA_DIR = paths.ANA_DIR
from greco_grb.scripts.utils import *
from glob import glob
import histlite as hl
import scipy

UNBLIND_PATH = "/storage/home/hhive1/cchen641/icecube/greco_grb/data/csky_output/unblind"
UNBLIND_RESULT = f"{UNBLIND_PATH}/result02.pkl"

result02 = pd.read_pickle(UNBLIND_RESULT)


result02.columns

# Unblind SoB at each TW
For each GRB, specify TW = 10, 25, 50, 100, 250, 500 and **prompt (T100)**, get the unblind events SoB

print("\n===== Setting up csky =====\n")
data_dir = ICDATA_DIR
data_filenames = sorted(glob(data_dir + '/IC86_20*.data.npy'))
sig_filenames = sorted(glob(data_dir + '/IC86_2012.nu*_merged.npy'))
grl_filenames = sorted(glob(data_dir + '/GRL/IC86_20*.data.npy'))

################ energy lower bound #############
min_log_e = np.log10(10)
#################################################
bins_sindec = np.linspace(-1, 1, 25+1)  
bins_logenergy = np.linspace(min_log_e, 4, 25+1)

data = [np.load(data_filename) for data_filename in data_filenames]
data = np.hstack(data)
sig = [np.load(sig_filename) for sig_filename in sig_filenames]
sig = np.hstack(sig)
grl = [np.load(grl_filename) for grl_filename in grl_filenames]
grl = np.hstack(grl)
if min_log_e is not None:
    data_mask = data['logE'] > min_log_e
    data = data[data_mask]
    sig_mask = sig['logE'] > min_log_e
    sig = sig[sig_mask]
    
dataset_spec = cy.selections.CustomDataSpecs.CustomDataSpec(data, sig, np.sum(grl['livetime']),
                                                     sindec_bins=bins_sindec,
                                                     logenergy_bins=bins_logenergy,
                                                     grl=grl, key='greco_v2.10', cascades=True)

ANA_DIR = cy.utils.ensure_dir(ANA_DIR)
# on OSG
# ana_dir = "./"
ana = cy.get_analysis(cy.selections.repo
                      , dataset_spec
                      , dir=ANA_DIR
                      , load_sig=True)  # to save memory: use False


class args:
    grb_idx_start = 0 
    grb_idx_end = 1 
    ncpu = 24         ################### Number of Cores #############
    out_dir = "/storage/home/hhive1/cchen641/data/icecube/data/greco_grb/data/csky_output"
############# used for spatial_prior_trial_runner
conf = {
    'ana': ana,
    #### llh basics: csky.conf
    'space': 'ps', # ps/fitps/template/prior
    'time': 'transient', # utf/lc/transient
    'energy': 'customflux', # fit/customflux
    'flux': cy.hyp.PowerLawFlux(2.5),
    #### inj.py - prior has some duplications against space's prior
    'sig': 'transient', # ps/tw/lc/transient/template/prior
    'full_sky': True,
    'extended': True,
    'mp_cpus': args.ncpu,
    'cut_n_sigma': 3,
}

cy.CONF.update(conf)

df = pd.read_pickle(DATA_DIR+"/grbwebgbm/grbweb_gbm_noHealpix_2268.pkl")

df

## 1. get s/b and n_b

## only uncomment below if want refreshed results

# sob_matrix = [
#     [],  # t100
#     [],  # 10 
#     [],  # 25
#     [],  # 50
#     [],  # 100
#     [],  # 250
#     [],  # 500
# ]
# n_b_list = [0,0,0,0,0,0,0]  # expect 1+6 values
# # for each grb
# # for grb_idx in range(args.grb_idx_start, args.grb_idx_end):
# timer = cy.timing.Timer()
# time = timer.time
# with time("Test..."):
#     end_idx_plus_1 = 2268 # 2268
#     for grb_idx in range(0, end_idx_plus_1):
#         # load grb info
#         grb_name = df.grb_name[grb_idx]
#         if(grb_idx % 20 == 0):
#             print(f"(report every 20) Working on {grb_name}")
#         grb_row = df.loc[df['grb_name'] == grb_name]
#         ra = grb_row.ra
#         dec = grb_row.dec
#         t100 = grb_row.t100 * 86400.
#         # load grb healpix map
#         healpix = np.load(DATA_DIR+"/grbwebgbm/healpix/{}_healpix_nside64.npy".format(grb_name))
#         healpix = np.maximum(healpix,0)
#         ########## healpix reduce (< instead of <=) ##########
#         healpix[healpix < isf_healpix(healpix, q=0.99)] = 0
#         healpix = healpix / np.sum(healpix)
#         # for each tw_in_second
#     #     tss, nss = [], []
#         for tw_idx, tw_in_second in enumerate([t100, 10,25,50,100,250,500]):
#             tw = tw_in_second/86400.
#             tw_start = grb_row.t_center - 0.5*tw
#             src = cy.sources(
#                 ra=ra,
#                 dec=dec,
#                 deg=True,
#                 mjd=tw_start, 
#                 sigma_t=np.zeros_like(tw), 
#                 t_100=tw,  # in days
#                 prior=[hl.heal.HealHist(healpix)],
#                 name=grb_name
#             )
#             tr = cy.get_trial_runner(conf=cy.CONF, ana=ana, src=src)
#             L = tr.get_one_llh(TRUTH=True)
#             #### i=0 for bg, i=1 for sig; unblind so i=0.
#             SB_space = cy.inspect.get_space_eval(L, -1, 0)()[0]
#             ## unblinded, no need for (gamma=xxx)
#             SB_energy = cy.inspect.get_energy_eval(L, -1, 0)()[0]
#             SB = SB_space * SB_energy
            
#             sob_matrix[tw_idx].append(SB)
#             n_b_list[tw_idx] += L.llh_model.N_bg
            
# with time("Saving..."):
#     np.save(DATA_DIR+f"/csky_output/comparison/unblind/{end_idx_plus_1}_sob_matrix.npy", 
#                 np.array(sob_matrix,dtype=object))
#     np.save(DATA_DIR+f"/csky_output/comparison/unblind/{end_idx_plus_1}_n_b_list.npy", 
#                 np.array(n_b_list,dtype=object))

## 2. get `n_fit` and `TS_unblind` from `s/b` and `n_b`

end_idx_plus_1 = 2268
sob_matrix_from_file = np.load(DATA_DIR+f"/csky_output/comparison/unblind/{end_idx_plus_1}_sob_matrix.npy", allow_pickle=True)
n_b_list_from_file = np.load(DATA_DIR+f"/csky_output/comparison/unblind/{end_idx_plus_1}_n_b_list.npy", allow_pickle=True)


sob_matrix_from_file, n_b_list_from_file

###########################
#                         #
#        Functions        #
#                         #
###########################


def neg_TS(n_s, sobe_list, n_b):
    """
    Calculate the test statistic for a given S/B list, 
        number of background events, and n_s.
        
    Define the function for use in a minimizer later.
    
    Returns the negative TS
    """
    TS = -n_s + np.sum(np.log(n_s * (sobe_list/n_b) +1))
    
    return -TS

def calc_stacked_TS(sobe_list, n_b):
    """
    Calculate the TS for a set of S/B and a stacked n_b.  
        This function calls a minimizer to find the n_s 
        that minimizes the negative TS value.
        
    sobe_list: numpy array
        Concatenated list of S/B values for all GRBs.  These 
        are the signal/background values for every event 
        contributing to a GRB in a trial
    n_b: float
        Number of expected background events for the entire 
        stacked time window
    
    Returns a test statistic and fit ns value for the trial.
    """
    
    # See the last cell of this script for several examples of how to make this work

    # A short function to sort out passed args versus what the minimizer should fit
    def fit_neg_ts(n_s): return neg_TS(n_s, sobe_list, n_b)
    
    # Fit the results with a scipy minimizer
    fit_results = scipy.optimize.minimize(fit_neg_ts, [1.], bounds=[(0.,len(sobe_list))]) 

    # Get the fit results
    ns_fit = fit_results.x[0]
    TS = -fit_results.fun[0]
    
    return TS, ns_fit



def calc_unblind_TS(sobe_list, n_b):
    """
    Calculate the unblinded TS for some subset of GRBs
    
    grb_results_array: numpy array
        Sky and duration cuts to determine which GRBs are considered
    TW_str: str
        Prompt or 0-9 for time window
        
    Parameter:
    ---------------
    sobe_list : numpy array
                the length of the list can be from 0 to a number greater than 2268
                
    n_b : float
            one number obtained from all 2268 GRBs
    """
    
    if len(sobe_list) != 0:
        sobe_list_combined = np.concatenate(sobe_list)
    else:
        sobe_list_combined = np.array([0.])

    #n_b = total_time*0.007  # Fixed background rate so calculate n_b easily
    
    TS_unblind, ns_fit = calc_stacked_TS(sobe_list_combined, n_b)
    if TS_unblind < 0:
        TS_unblind = 0.
        
    return TS_unblind, ns_fit

sob_matrix_from_file.shape, n_b_list_from_file.shape

# [t100, 10,25,50,100,250,500]
[calc_unblind_TS(sob_matrix_from_file[i], n_b_list_from_file[i]) for i in range(0, 7)]

# 3. Get sob, n_b and stacked TS from background runs

class args:
    grb_name = "GRB180423A"    # real healpix example
    # grb_name = "GRB190415A"    # fake healpix example 
    batchNtrials = 5
    batchIndex = 0
    ncpu = 24
    outfilename = ""

rng=np.random.default_rng(abs(java_hash(args.grb_name)))
seeds = rng.integers(int(1e9), size=int(2e8))[args.batchNtrials*args.batchIndex: args.batchNtrials*(args.batchIndex + 1)]
print("\n...Done\n")

seeds

grb_row.t100.values[0]

grb_name = args.grb_name
grb_row = df.loc[df['grb_name'] == grb_name]
ra = grb_row.ra
dec = grb_row.dec
t100 = grb_row.t100.values[0] * 86400.
### load grb healpix map
healpix = np.load(DATA_DIR+"/grbwebgbm/healpix/{}_healpix_nside64.npy".format(grb_name))
healpix = np.maximum(healpix,0)
########## healpix reduce (< instead of <=) ##########
healpix[healpix < isf_healpix(healpix, q=0.99)] = 0
healpix = healpix / np.sum(healpix)
### for each tw_in_second
for tw_idx, tw_in_second in enumerate([t100, 10,25,50,100,250,500]):
    timer = cy.timing.Timer()
    time = timer.time
    with time(f"Working on {grb_name} with TW{tw_in_second}..."):
        sob_list = []
        n_b_list =[]
        tw = tw_in_second/86400.
        tw_start = grb_row.t_center - 0.5*tw
        src = cy.sources(
            ra=ra,
            dec=dec,
            deg=True,
            mjd=tw_start, 
            sigma_t=np.zeros_like(tw), 
            t_100=tw,  # in days
            prior=[hl.heal.HealHist(healpix)],
            name=grb_name
        )
        for no_trial in range(args.batchNtrials):
            sptr = cy.get_spatial_prior_trial_runner(conf=cy.CONF
                                         ,src_tr=src
                                         ,llh_priors=[healpix])

            ############## ref: osg_stacked_bg_tsd.py ########################
            ############## Get the pixelmax ##################################
            try:
                _, ts, ns, ra_fit, dec_fit = sptr.get_one_fit(TRUTH=False, 
                                                               mp_cpus=args.ncpu, 
                                                               logging=False,
                                                              seed=seeds[no_trial])
            except:
                ts, ns, ra_fit, dec_fit = 0.0, 0.0, 0.0, 0.0
            if ns == 0.0:
                ra_fit, dec_fit = grb_row.ra, grb_row.dec
            ############## Get SoB, n_b for point source: pixelmax ##########
            src_fit = cy.sources(
                            ra=ra_fit,
                            dec=dec_fit,
                            deg=True,
                            mjd=tw_start, 
                            sigma_t=np.zeros_like(tw), 
                            t_100=tw,  # in days
                            # prior=[hl.heal.HealHist(healpix)], # not important here
                            name=grb_name
                        )
            tr = cy.get_trial_runner(conf=cy.CONF, ana=ana, src=src_fit)
            L = tr.get_one_llh(TRUTH=False, seed=seeds[no_trial])
            try:
                #### i=0 for bg, i=1 for sig; unblind so i=0.
                SB_space = cy.inspect.get_space_eval(L, -1, 0)()[0]
                ## unblinded, no need for (gamma=xxx)
                SB_energy = cy.inspect.get_energy_eval(L, -1, 0)()[0]
                SB = SB_space * SB_energy
            except:
                SB = np.array([])
            n_b = L.llh_model.N_bg
            sob_list.append(SB)
            n_b_list.append(n_b)
    # Save sob_list and n_b_list for this GRB with this TW
    print(f"Saving SoB and n_b for {grb_name} with TW={tw_in_second:.2f}")
    output_path = cy.utils.ensure_dir(DATA_DIR+f"/csky_output/comparison/background/{grb_name}")
    np.save(f"{output_path}/{grb_name}_tw{tw_in_second:.2f}_batchIndex{args.batchIndex}_batchNtrials{args.batchNtrials}.npy", np.array(sob_list))
    np.save(f"{output_path}/{grb_name}_tw{tw_in_second:.2f}_batchIndex{args.batchIndex}_batchNtrials{args.batchNtrials}.npy", np.array(n_b_list))

n_b_list

# 4. Calculate TS_upper_limit based on TS_median from background and TS_unblind above
